ClothSegmenter/pyproject.toml
[tool.poetry]
name = "garmentsegmenter"
version = "0.1.0"
description = "FastAPI serving from hugging face API."
authors = ["rlrahulkanojia"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
gdown = "^5.2.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
pillow = "^11.1.0"
python-json-logger = "^3.2.1"
gunicorn = "~22.0"
requests = "~2.31"
uvicorn = {version = "~0", extras = ["standard"]}
pydantic = "^2.10.5"
python-multipart = "^0.0.20"


[tool.poetry.group.infer.dependencies]
opencv-python = "^4.11.0.86"
ipykernel = "^6.29.5"
mypy = "^1.14.1"
streamlit = "^1.41.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
